+++ src/DependencyInjection/Configuration.php
@@ -27,10 +27,6 @@ class Configuration implements ConfigurationInterface
                     ->info('Api-Key value given on registration of the user.')
                     ->isRequired()
                 ->end()
-                ->scalarNode('api_secret')
-                    ->info('Deprecated. This is not used anymore.')
-                    ->defaultNull()
-                ->end()
                 ->scalarNode('organization')
                     ->info('Organization value like "my-org" for use in calls to the API.')
                     ->isRequired()
@@ -40,6 +36,14 @@ class Configuration implements ConfigurationInterface
                     ->defaultValue('https://api.rokka.io')
                 ->end()
             ->end();
+
+        $secretNode = $rootNode->children()->scalarNode('api_secret')
+            ->info('Deprecated. This is not used anymore.')
+            ->defaultNull();
+
+        if (method_exists($secretNode, 'setDeprecated')) {
+            $secretNode->setDeprecated();
+        }
 
         return $treeBuilder;
     }