+++ tests/Functional/Fixtures/AppKernel.php
@@ -5,9 +5,6 @@ namespace Rokka\RokkaClientBundle\Tests\Functional\Fixtures;
 use Symfony\Component\Config\Loader\LoaderInterface;
 use Symfony\Component\HttpKernel\Kernel;
 
-/* BC to run tests on Symfony < 3.3 */
-class_alias(AppKernel::class, '\AppKernel');
-
 class AppKernel extends Kernel
 {
     public function registerBundles()
@@ -33,3 +30,6 @@ class AppKernel extends Kernel
         return sys_get_temp_dir().'/rokka-client-bundle/logs';
     }
 }
+
+/* BC to run tests on Symfony < 3.3 */
+class_alias(AppKernel::class, \AppKernel::class, false);