+++ changelog.md
@@ -12,11 +12,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Removed
 ### Security
 
-## [1.1.0] - 2017-12-06
+## [1.1.0] - 2017-12-08
 ### Added
 - Symfony 4 support
 ### Changed
 - Made api_secret optional in configuration, since it's not used anymore.
+- Dropped Symfony 3.0 and 3.1 support (Symfony 2 LTS still supported)
 
 ## [1.0.1] - 2017-11-29
 ### Fixed